backend/requirements.txt
fastapi==0.104.1
<<<<<<< HEAD
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
=======
sqlmodel==0.0.16
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose==3.3.0
uvicorn==0.24.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
>>>>>>> 3472c09df4920281bb81456aab4e032e839461c6
